refactor: 카테고리 조회 api 호출 위치 변경
